Rename files from EBonsai to FLOWER
